+++ packages/sdks/src/blocks/button/button.types.ts
@@ -1,0 +1,6 @@
+export interface ButtonProps {
+  attributes?: any;
+  text?: string;
+  link?: string;
+  openLinkInNewTab?: boolean;
+}
