+++ packages/sdks/src/blocks/fragment/fragment.lite.tsx
@@ -1,15 +1,11 @@
 import { useMetadata } from '@builder.io/mitosis';
+import type { FragmentProps } from './fragment.types';
 
 useMetadata({
   rsc: {
     componentType: 'client',
   },
 });
-export interface FragmentProps {
-  maxWidth?: number;
-  attributes?: any;
-  children?: any;
-}
 
 export default function FragmentComponent(props: FragmentProps) {
   // TODO: flag for if target supports fragments / doesn't need root/host elements
