+++ packages/sdks/src/blocks/fragment/fragment.types.ts
@@ -1,0 +1,5 @@
+export interface FragmentProps {
+  maxWidth?: number;
+  attributes?: any;
+  children?: any;
+}
