+++ packages/sdks/src/blocks/image/image.types.ts
@@ -1,0 +1,20 @@
+import type { BuilderBlock } from '../../types/builder-block';
+
+export interface ImageProps {
+  className?: string;
+  image: string;
+  sizes?: string;
+  lazy?: boolean;
+  height?: number;
+  width?: number;
+  altText?: string;
+  backgroundSize?: 'cover' | 'contain';
+  backgroundPosition?: string;
+  srcset?: string;
+  aspectRatio?: number;
+  children?: any;
+  fitContent?: boolean;
+  builderBlock?: BuilderBlock;
+  noWebp?: boolean;
+  src?: string;
+}
