+++ packages/sdks/src/blocks/section/section.types.ts
@@ -1,0 +1,6 @@
+export interface SectionProps {
+  maxWidth?: number;
+  attributes?: any;
+  children?: any;
+  builderBlock?: any;
+}
