+++ packages/sdks/src/blocks/symbol/symbol.types.ts
@@ -1,0 +1,19 @@
+import type { BuilderContent } from '../../types/builder-content';
+import type { BuilderComponentsProp } from '../../types/builder-props';
+
+interface SymbolInfo {
+  model?: string;
+  entry?: string;
+  data?: any;
+  content?: BuilderContent;
+  inline?: boolean;
+  dynamic?: boolean;
+}
+
+export interface SymbolProps extends BuilderComponentsProp {
+  symbol?: SymbolInfo;
+  dataOnly?: boolean;
+  dynamic?: boolean;
+  attributes?: any;
+  inheritState?: boolean;
+}
