+++ packages/sdks/src/blocks/text/text.types.ts
@@ -1,0 +1,3 @@
+export type TextProps = {
+  text?: string;
+};
