+++ packages/sdks/src/blocks/video/video.types.ts
@@ -1,0 +1,31 @@
+import type { BuilderBlock } from '../../types/builder-block';
+
+export interface VideoProps {
+  attributes?: any;
+  video?: string;
+  autoPlay?: boolean;
+  controls?: boolean;
+  muted?: boolean;
+  loop?: boolean;
+  playsInline?: boolean;
+  aspectRatio?: number;
+  width?: number;
+  height?: number;
+  fit?: 'contain' | 'cover' | 'fill';
+  preload?: 'auto' | 'metadata' | 'none';
+  position?:
+    | 'center'
+    | 'top'
+    | 'left'
+    | 'right'
+    | 'bottom'
+    | 'top left'
+    | 'top right'
+    | 'bottom left'
+    | 'bottom right';
+  posterImage?: string;
+  lazyLoad?: boolean;
+  children?: any;
+  fitContent?: boolean;
+  builderBlock?: BuilderBlock;
+}
