+++ packages/sdks/src/server-index.ts
@@ -1,5 +1,16 @@
 export * from './index-helpers/top-of-file.js';
 
+export { ButtonProps } from './blocks/button/button.types.js';
+export { ColumnProps } from './blocks/columns/columns.types.js';
+export { FragmentProps } from './blocks/fragment/fragment.types.js';
+export { ImageProps } from './blocks/image/image.types.js';
+export { SectionProps } from './blocks/section/section.types.js';
+export { SymbolProps } from './blocks/symbol/symbol.types.js';
+export { TextProps } from './blocks/text/text.types.js';
+export { VideoProps } from './blocks/video/video.types.js';
+export { BlocksProps } from './components/blocks/blocks.types.js';
+export { ContentVariantsPrps as ContentProps } from './components/content-variants/content-variants.types.js';
+
 export { isEditing } from './functions/is-editing.js';
 export { isPreviewing } from './functions/is-previewing.js';
 export { createRegisterComponentMessage } from './functions/register-component.js';
@@ -22,7 +33,6 @@ export { getBuilderSearchParams } from './functions/get-builder-search-params/in
 
 export { track } from './functions/track/index.js';
 
-export { ContentVariantsPrps as ContentProps } from './components/content-variants/content-variants.types.js';
 export type { RegisteredComponent } from './context/types.js';
 export type { ComponentInfo } from './types/components.js';
